Refactor exceptions into exception package
