+++ src/main/java/Hamlet.java
@@ -1,9 +1,9 @@
 package main.java;
 
-import java.util.ArrayList;
+import main.java.exception.HamletException;
+
 import java.util.Scanner;
 import java.io.IOException;
-import java.time.format.DateTimeFormatter;
 import java.time.format.DateTimeParseException;
