+++ src/main/java/Parser.java
@@ -1,5 +1,9 @@
 package main.java;
 
+import main.java.exception.DeadlineException;
+import main.java.exception.EventException;
+import main.java.exception.TodoException;
+
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
 import java.util.ArrayList;
