+++ src/main/java/exception/DeadlineException.java
@@ -1,6 +1,6 @@
-package main.java;
+package main.java.exception;
 
-public class DeadlineException extends HamletException{
+public class DeadlineException extends HamletException {
     @Override
     public String toString() {
         return "The description or /by of a deadline cannot be empty.";
